Project: rename eXplorer project to avoid collision with explorer.exe on Windows.
The generated executable was called 'eXplorer.exe'. When attempting to attach Renderdoc, it would launch the executable and then silently detach. The reason was that on Windows, Renderdoc explicitly detects attempts to attach to 'explorer.exe' (case insensitive) and prevents them. A warning message is output _from the injected instance of the DLL_, which was not redirected to the main log file.